Fixes doc generation and service worker matching patterns
Previously, directories/files in _pages/ that were prefixed with
the '_' symbol (e.g. '_pages/_guides/') were being built and
copied to the 'pages/' output directory. These files were used
to hold content not ready for publication. This is a standard
way of marking things that should not be output into the built
files in static site generators like Jekyll.

After the files were copied to the 'pages/' directory, the service
worker generate gulp task would add those '_' files to the list
for the browser to cache.

Then, those files would NOT get pushed to production because a
gitignore pattern caused them to be ignored when the production
build committed all the files.

This led to the service worker attempting to fetch these files
when they did not exist on production. If a service worker
pre-caching files hits a 404, it bails and cancels registration,
causing a performance hit and users to not get the offline cache.

This commit fixes this bug by ensuring no '_' prefixed files or
directories will be copied from '_pages/' during the build.